First commit for collection report parser

+++ mapping/src/main/java/com/mapping/parser/input/TrimatrixTracker.java
@@ -124,7 +124,7 @@ public class TrimatrixTracker implements Serializable {
 		this.invoiceNumber = invoiceNumber;
 	}
 
-	public double getReceiptNumber() {
+	public long getReceiptNumber() {
 		return receiptNumber;
 	}
